First pass at recursive mode.
If there's exactly one function/constructor call in the definitions
found, find another definition by its name.

I'm not totally sure I like this criterion: if you happen to have a
small class that only has one method that even does a function call,
how likely are you to be interested in the definition of the called
function?