Follow imports in python.
When we see clues that the symbol we're looking for is from some
import path, reorder the remaining files to first search likely
matches of the import statement.

Costs maybe an extra 50% time so switch to printing results as we find
them.

Include `--verbose` mode, mostly to help me verify that imports are
doing anything.