Add LaTeX support.
Requires the tree-sitter CLI since the parser maintainers opted not to
check in the 40-meg generated parser.c.

Includes a bit of a hack to strip initial backslashes. Otherwise you'd
have to type up to 4 backslashes at the beginning of a search---which
become 2 after shell escaping and then 1 after regex escaping.